fix: vLLM E2E 2740/2740 -- arch remap, force_flush, final_logits slicing
- Remap HF architectures to hooked models (GPT2PLMHeadModel,
  Qwen3PForCausalLM) in load_model before super() so vLLM registry
  resolves the correct model class with HookPoints
- RingEngine::stop() calls cudaDeviceSynchronize + force_flush_and_wait
  before stopping drain/p2p threads, ensuring all entries are flushed
- Fix final_logits slicing in p2p_thread: vLLM compute_logits returns
  [num_reqs, vocab] indexed by request position, not token offset.
  DB token range adjusted to (end_token-1, end_token) for both HF and
  vLLM paths based on actual slice dim0
- HookPoint.forward: change from-import to module-level access for
  _active_transport (cpu_direct path preserved, functionally identical)
- Read token_ids dtype from model_runner buffer (vLLM) or input_ids
  tensor (HF) instead of hardcoding torch.int32
- Null mode on during warmup, off after compile_or_warm_up_model
- Flattened mode: batch=0, logits_to_keep=num_reqs

+++ monitoring/generate.py
@@ -399,9 +399,13 @@ def _install_prepare_wrapper(model: Any) -> None:
                         if _profile:
                             _t_prepare = _time.perf_counter()
 
-                    # Push FIFO metadata for p2p thread
+                    # Push FIFO metadata for p2p thread.
+                    # input_ids_val is guaranteed non-None here (guarded above).
+                    # On non-first PP ranks input_ids would be None and we
+                    # skip this entire block, so .dtype is always safe.
                     transport.pre_push_all_metas(batch, q_len, kv_dim,
-                                                logits_to_keep=logits_to_keep)
+                                                logits_to_keep=logits_to_keep,
+                                                token_ids_dtype=input_ids_val.dtype)
 
                     if _profile:
                         _t_meta = _time.perf_counter()
